Added regex filter to ensure callback is valid JavaScript function.
This will overcome an XSS vulnerability that allows any string data
inside the callback parameter - as long as the data is a string. This
filters the callback parmeter to verify it only contains alpha (also
capital), numeric characters, underscore, and dollar sign ($), also
ensuring the
begining is not a number.